Move `UserScriptSettings` and `OpeningBehaviorSettings` to `plugins.settings` package
