+++ src/main/java/sc/fiji/ome/zarr/plugins/settings/OpeningBehaviorSettings.java
@@ -26,7 +26,7 @@
  * POSSIBILITY OF SUCH DAMAGE.
  * #L%
  */
-package sc.fiji.ome.zarr.plugins;
+package sc.fiji.ome.zarr.plugins.settings;
 
 import java.lang.invoke.MethodHandles;
 import java.util.Arrays;
