+++ src/test/java/sc/fiji/ome/zarr/plugins/settings/OpeningBehaviorSettingsTest.java
@@ -26,7 +26,7 @@
  * POSSIBILITY OF SUCH DAMAGE.
  * #L%
  */
-package sc.fiji.ome.zarr.plugins;
+package sc.fiji.ome.zarr.plugins.settings;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
